added timer to fix refresh

+++ src/GUI/UI/ExperimentResults/ExperimentListPanel.py
@@ -28,10 +28,9 @@ class ExperimentListPanel(SpaeDisplayPanel):
         # Adds all the experiments in the application queue to the display list
 
         self.root = None
-        self.reload(None)
 
-        self.run_button = wx.Button(self)
-        self.run_button.SetLabelText("Reload the Results")
+        # self.run_button = wx.Button(self)
+        # self.run_button.SetLabelText("Reload the Results")
 
 
         # Runs deselected on a double click or when escape is pressed
@@ -41,11 +40,18 @@ class ExperimentListPanel(SpaeDisplayPanel):
         self.sizer = wx.BoxSizer(wx.VERTICAL)
         self.SetSizer(self.sizer)
         self.sizer.Add(self.tree_box, 5, wx.EXPAND | wx.ALL)
-        self.sizer.Add(self.run_button, 1, wx.EXPAND | wx.ALL)
+        # self.sizer.Add(self.run_button, 1, wx.EXPAND | wx.ALL)
 
         # Runs the selected function when an experiment is selected
         self.Bind(wx.EVT_TREE_SEL_CHANGED, self.selected)
         self.Bind(wx.EVT_BUTTON, self.reload)
+        # self.Bind(wx.EVT_TIMER, self.reload, self.Parent.Parent.Parent.Parent.timer)
+
+        self.timer = wx.Timer(self)
+        self.Bind(wx.EVT_TIMER, self.reload, self.timer)
+        self.timer.Start(5000)
+
+        self.reload(None)
 
     def set_up_ui_control(self, ui_control):
         ui_control.add_control_to_text_list(self.run_button)
@@ -54,14 +60,18 @@ class ExperimentListPanel(SpaeDisplayPanel):
         """
         Reloads it self, updates if the Queue has changed
         """
+        print "RELOADING", self.root, self, event
         if not self.root:
             self.root = self.tree_box.AddRoot(GUI_CONSTANTS.EXPERIMENT_QUEUE_RESULT_ROOT)
-        self.tree_box.DeleteChildren(self.root)
-        for que_result in Globals.systemConfigManager.get_results_manager().get_queue_results():
-            que_root = self.tree_box.AppendItem(self.root, que_result.get_name())
-            for exp_result_name in que_result.get_experiment_results_list():
-                self.tree_box.AppendItem(que_root, exp_result_name)
-        self.tree_box.ExpandAll()
+        # print self.tree_box.GetChildrenCount(self.root, recursively=False), len(Globals.systemConfigManager.get_results_manager().get_queue_results())
+        if self.tree_box.GetChildrenCount(self.root, recursively=False) != len(Globals.systemConfigManager.get_results_manager().get_queue_results()):
+
+            self.tree_box.DeleteChildren(self.root)
+            for que_result in Globals.systemConfigManager.get_results_manager().get_queue_results():
+                que_root = self.tree_box.AppendItem(self.root, que_result.get_name())
+                for exp_result_name in que_result.get_experiment_results_list():
+                    self.tree_box.AppendItem(que_root, exp_result_name)
+            self.tree_box.ExpandAll()
 
     def deselected(self, event):
         """
